Remove 'trim' dependency
The 'trim' dependency is problematic as it doesn't link to its source
code from the package.json and while it mentions in README it
is licensed under MIT, it doesn't have a full license text in
the npm package distribution. The MIT license explicitly specifies:

"The above copyright notice and this permission notice shall be
included in all copies or substantial portions of the Software."

Thus not attaching the full license text is a discrepancy with
the license itself. That said, the main reason to remove the dependency
is that it should not be necessary anymore. The ES5 includes
String.prototype.trim() method and according to MDN (https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/Trim)
it is fairly well supported even in browsers.